confint.fracdiff {fracdiff}	R Documentation

Confidence Intervals for Fracdiff Model Parameters

Description

Computes confidence intervals for one or more parameters in a fitted fracdiff model, see fracdiff.

Usage

## S3 method for class 'fracdiff'
confint(object, parm, level = 0.95, ...)

Arguments

object
an object of class fracdiff, typically result of fracdiff(..).

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the confidence level required.

...
additional argument(s) for methods.

Value

A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).

Author(s)

Spencer Graves posted the initial version to R-help.

See Also

the generic confint; fracdiff model fitting.

Examples

set.seed(101)
ts2 <- fracdiff.sim(5000, ar = .2, ma = -.4, d = .3)
mFD <- fracdiff( ts2$series, nar = length(ts2$ar), nma = length(ts2$ma))
coef(mFD)
confint(mFD)

[Package fracdiff version 1.4-2 Index]