diffseries {fracdiff}	R Documentation

Fractionally Differenciate Data

Description

Differenciates the time series data using the approximated binomial expression of the long-memory filter and an estimate of the memory parameter in the ARFIMA(p,d,q) model.

Usage

diffseries(x, d)

Arguments

x
numeric vector or univariate time series

.

d
number specifiying the fractional difference order.

Value

the fractionally differenced series x.

Author(s)

Valderio A. Reisen valderio@cce.ufes.br and Artur J. Lemonte

References

See those in fdSperio; additionally

Reisen, V. A. and Lopes, S. (1999) Some simulations and applications of forecasting long-memory time series models; Journal of Statistical Planning and Inference 80, 269–287.

Reisen, V. A. Cribari-Neto, F. and Jensen, M.J. (2003) Long Memory Inflationary Dynamics. The case of Brazil. Studies in Nonlinear Dynamics and Econometrics 7(3), 1–16.

See Also

fracdiff.sim

Examples

memory.long <- fracdiff.sim(80, d = 0.3)
mGPH <- fdGPH(memory.long$series)
r <- diffseries(memory.long$series, d = mGPH$d)
#acf(r) # shouldn't show structure - ideally

[Package fracdiff version 1.4-2 Index]