fracdiff-methods {fracdiff}	R Documentation

Many Methods for "fracdiff" Objects

Description

Many “accessor” methods for fracdiff objects, notably summary, coef, vcov, and logLik; further print() methods were needed.

Usage

## S3 method for class 'fracdiff'
coef(object, ...)
## S3 method for class 'fracdiff'
logLik(object, ...)
## S3 method for class 'fracdiff'
print(x, digits = getOption("digits"), ...)
## S3 method for class 'fracdiff'
summary(object, symbolic.cor = FALSE, ...)
## S3 method for class 'summary.fracdiff'
print(x, digits = max(3, getOption("digits") - 3),
        correlation = FALSE, symbolic.cor = x$symbolic.cor,
        signif.stars = getOption("show.signif.stars"), ...)

## S3 method for class 'fracdiff'
vcov(object, ...)

Arguments

x, object
object of class fracdiff.

digits
the number of significant digits to use when printing.

...
further arguments passed from and to methods.

correlation
logical; if TRUE, the correlation matrix of the estimated parameters is returned and printed.

symbolic.cor
logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.

signif.stars
logical. If TRUE, “significance stars” are printed for each coefficient.

Author(s)

Martin Maechler

See Also

fracdiff to get "fracdiff" objects, confint.fracdiff for the confint method; further, fracdiff.var.

Examples

set.seed(7)
ts4 <- fracdiff.sim(10000, ar = c(0.6, -.05, -0.2), ma = -0.4, d = 0.2)
modFD <- fracdiff( ts4$series, nar = length(ts4$ar), nma = length(ts4$ma))
## -> warning (singular Hessian) %% FIXME ???
coef(modFD) # the estimated parameters
vcov(modFD)
smFD <- summary(modFD)
smFD
coef(smFD) # gives the whole table
AIC(modFD) # AIC works because of the logLik() method

[Package fracdiff version 1.4-2 Index]