fdGPH {fracdiff}	R Documentation

Geweke and Porter-Hudak Estimator for ARFIMA(p,d,q)

Description

Estimate the fractional (or “memory”) parameter d in the ARFIMA(p,d,q) model by the method of Geweke and Porter-Hudak (GPH). The GPH estimator is based on the regression equation using the periodogram function as an estimate of the spectral density.

Usage

fdGPH(x, bandw.exp = 0.5)

Arguments

x
univariate time series

bandw.exp
the bandwidth used in the regression equation

Details

The function also provides the asymptotic standard deviation and the standard error deviation of the fractional estimator.

The bandwidth is bw = trunc(n ^ bandw.exp), where 0 < bandw.exp < 1 and n is the sample size. Default bandw.exp = 0.5.

Value

d
GPH estimate

sd.as
asymptotic standard deviation

sd.reg
standard error deviation

Author(s)

Valderio A. Reisen and Artur J. Lemonte

References

see those in fdSperio.

See Also

fdSperio, fracdiff

Examples

memory.long <- fracdiff.sim(1500, d = 0.3)
fdGPH(memory.long$series)

[Package fracdiff version 1.4-2 Index]