fdSperio {fracdiff}	R Documentation

Sperio Estimate for 'd' in ARFIMA(p,d,q)

Description

This function makes use Reisen (1994) estimator to estimate the memory parameter d in the ARFIMA(p,d,q) model. It is based on the regression equation using the smoothed periodogram function as an estimate of the spectral density.

Usage

fdSperio(x, bandw.exp = 0.5, beta = 0.9)

Arguments

x
univariate time series data.

bandw.exp
numeric: exponent of the bandwidth used in the regression equation.

beta
numeric: exponent of the bandwidth used in the lag Parzen window.

Details

The function also provides the asymptotic standard deviation and the standard error deviation of the fractional estimator.

The bandwidths are bw = trunc(n ^ bandw.exp), where 0 < bandw.exp < 1 and n is the sample size. Default bandw.exp= 0.5;
and bw2 = trunc(n ^ beta), where 0 < beta < 1 and n is the sample size. Default beta = 0.9.

Value

a list with components

d
Sperio estimate

sd.as
asymptotic standard deviation

sd.reg
standard error deviation

Author(s)

Valderio A. Reisen valderio@cce.ufes.br and Artur J. Lemonte

References

Geweke, J. and Porter-Hudak, S. (1983) The estimation and application of long memory time series models. Journal of Time Series Analysis 4(4), 221–238.

Reisen, V. A. (1994) Estimation of the fractional difference parameter in the ARFIMA(p,d,q) model using the smoothed periodogram. Journal Time Series Analysis, 15(1), 335–350.

Reisen, V. A., B. Abraham, and E. M. M. Toscano (2001) Parametric and semiparametric estimations of stationary univariate ARFIMA model. Brazilian Journal of Probability and Statistics 14, 185–206.

See Also

fdGPH, fracdiff

Examples

memory.long <- fracdiff.sim(1500, d = 0.3)
spm <- fdSperio(memory.long$series)
str(spm, digits=6)

[Package fracdiff version 1.4-2 Index]